# docs/api/swagger.yaml
basePath: /
definitions:
...
    account:
        description: The modelled account can be either a remote account, or one on this instance.
        properties:
            acct:
                description: |-
                    The account URI as discovered via webfinger.
                    Equal to username for local users, or username@domain for remote users.
                example: some_user@example.org
                type: string
                x-go-name: Acct
            avatar:
                description: Web location of the account's avatar.
                example: https://example.org/media/some_user/avatar/original/avatar.jpeg
                type: string
                x-go-name: Avatar
            avatar_static:
                description: |-
                    Web location of a static version of the account's avatar.
                    Only relevant when the account's main avatar is a video or a gif.
                example: https://example.org/media/some_user/avatar/static/avatar.png
                type: string
                x-go-name: AvatarStatic
            bot:
                description: Account identifies as a bot.
                type: boolean
                x-go-name: Bot
            created_at:
                description: When the account was created (ISO 8601 Datetime).
                example: "2021-07-30T09:20:25+00:00"
                type: string
                x-go-name: CreatedAt
            custom_css:
                description: CustomCSS to include when rendering this account's profile or statuses.
                type: string
                x-go-name: CustomCSS
            discoverable:
                description: Account has opted into discovery features.
                type: boolean
                x-go-name: Discoverable
            display_name:
                description: The account's display name.
                example: big jeff (he/him)
                type: string
                x-go-name: DisplayName
            emojis:
                description: Array of custom emojis used in this account's note or display name.
                items:
                    $ref: '#/definitions/emoji'
                type: array
                x-go-name: Emojis
            enable_rss:
                description: Account has enabled RSS feed.
                type: boolean
                x-go-name: EnableRSS
            fields:
                description: Additional metadata attached to this account's profile.
                items:
                    $ref: '#/definitions/field'
                type: array
                x-go-name: Fields
            followers_count:
                description: Number of accounts following this account, according to our instance.
                format: int64
                type: integer
                x-go-name: FollowersCount
            following_count:
                description: Number of account's followed by this account, according to our instance.
                format: int64
                type: integer
                x-go-name: FollowingCount
            header:
                description: Web location of the account's header image.
                example: https://example.org/media/some_user/header/original/header.jpeg
                type: string
                x-go-name: Header
            header_static:
                description: |-
                    Web location of a static version of the account's header.
                    Only relevant when the account's main header is a video or a gif.
                example: https://example.org/media/some_user/header/static/header.png
                type: string
                x-go-name: HeaderStatic
            id:
                description: The account id.
                example: 01FBVD42CQ3ZEEVMW180SBX03B
                type: string
                x-go-name: ID
            last_status_at:
                description: When the account's most recent status was posted (ISO 8601 Datetime).
                example: "2021-07-30T09:20:25+00:00"
                type: string
                x-go-name: LastStatusAt
            locked:
                description: Account manually approves follow requests.
                type: boolean
                x-go-name: Locked
            mute_expires_at:
                description: If this account has been muted, when will the mute expire (ISO 8601 Datetime).
                example: "2021-07-30T09:20:25+00:00"
                type: string
                x-go-name: MuteExpiresAt
            note:
                description: Bio/description of this account.
                type: string
                x-go-name: Note
            role:
                $ref: '#/definitions/accountRole'
            source:
                $ref: '#/definitions/Source'
            statuses_count:
                description: Number of statuses posted by this account, according to our instance.
                format: int64
                type: integer
                x-go-name: StatusesCount
            suspended:
                description: Account has been suspended by our instance.
                type: boolean
                x-go-name: Suspended
            url:
                description: Web location of the account's profile page.
                example: https://example.org/@some_user
                type: string
                x-go-name: URL
            username:
                description: The username of the account, not including domain.
                example: some_user
                type: string
                x-go-name: Username
        title: Account models a fediverse account.
        type: object
        x-go-name: Account
        x-go-package: github.com/superseriousbusiness/gotosocial/internal/api/model
...
    instanceConfigurationAccounts:
        properties:
            allow_custom_css:
                description: Whether or not accounts on this instance are allowed to upload custom CSS for profiles and statuses.
                example: false
                type: boolean
                x-go-name: AllowCustomCSS
            max_featured_tags:
                description: |-
                    The maximum number of featured tags allowed for each account.
                    Currently not implemented, so this is hardcoded to 10.
                format: int64
                type: integer
                x-go-name: MaxFeaturedTags
        title: InstanceConfigurationAccounts models instance account config parameters.
        type: object
        x-go-name: InstanceConfigurationAccounts
        x-go-package: github.com/superseriousbusiness/gotosocial/internal/api/model
...
    instanceV1:
        properties:
            account_domain:
                description: |-
                    The domain of accounts on this instance.
                    This will not necessarily be the same as
                    simply the Host part of the URI.
                example: example.org
                type: string
                x-go-name: AccountDomain
            approval_required:
                description: New account registrations require admin approval.
                type: boolean
                x-go-name: ApprovalRequired
            configuration:
                $ref: '#/definitions/instanceV1Configuration'
            contact_account:
                $ref: '#/definitions/account'
            description:
                description: |-
                    Description of the instance.

                    Should be HTML formatted, but might be plaintext.

                    This should be displayed on the 'about' page for an instance.
                type: string
                x-go-name: Description
            email:
                description: An email address that may be used for inquiries.
                example: admin@example.org
                type: string
                x-go-name: Email
            invites_enabled:
                description: Invites are enabled on this instance.
                type: boolean
                x-go-name: InvitesEnabled
            languages:
                description: Primary language of the instance.
                example: en
                items:
                    type: string
                type: array
                x-go-name: Languages
            max_toot_chars:
                description: |-
                    Maximum allowed length of a post on this instance, in characters.

                    This is provided for compatibility with Tusky and other apps.
                example: 5000
                format: uint64
                type: integer
                x-go-name: MaxTootChars
            registrations:
                description: New account registrations are enabled on this instance.
                type: boolean
                x-go-name: Registrations
            short_description:
                description: |-
                    A shorter description of the instance.

                    Should be HTML formatted, but might be plaintext.

                    This should be displayed on the instance splash/landing page.
                type: string
                x-go-name: ShortDescription
            stats:
                additionalProperties:
                    format: int64
                    type: integer
                description: 'Statistics about the instance: number of posts, accounts, etc.'
                type: object
                x-go-name: Stats
            thumbnail:
                description: URL of the instance avatar/banner image.
                example: https://example.org/files/instance/thumbnail.jpeg
                type: string
                x-go-name: Thumbnail
            thumbnail_description:
                description: Description of the instance thumbnail.
                example: picture of a cute lil' friendly sloth
                type: string
                x-go-name: ThumbnailDescription
            thumbnail_type:
                description: MIME type of the instance thumbnail.
                example: image/png
                type: string
                x-go-name: ThumbnailType
            title:
                description: The title of the instance.
                example: GoToSocial Example Instance
                type: string
                x-go-name: Title
            uri:
                description: The URI of the instance.
                example: https://gts.example.org
                type: string
                x-go-name: URI
            urls:
                $ref: '#/definitions/instanceV1URLs'
            version:
                description: |-
                    The version of GoToSocial installed on the instance.

                    This will contain at least a semantic version number.

                    It may also contain, after a space, the short git commit ID of the running software.
                example: 0.1.1 cb85f65
                type: string
                x-go-name: Version
        title: InstanceV1 models information about this instance.
        type: object
        x-go-name: InstanceV1
        x-go-package: github.com/superseriousbusiness/gotosocial/internal/api/model
    instanceV1Configuration:
        properties:
            accounts:
                $ref: '#/definitions/instanceConfigurationAccounts'
            emojis:
                $ref: '#/definitions/InstanceConfigurationEmojis'
            media_attachments:
                $ref: '#/definitions/instanceConfigurationMediaAttachments'
            polls:
                $ref: '#/definitions/instanceConfigurationPolls'
            statuses:
                $ref: '#/definitions/instanceConfigurationStatuses'
        title: InstanceV1Configuration models instance configuration parameters.
        type: object
        x-go-name: InstanceV1Configuration
        x-go-package: github.com/superseriousbusiness/gotosocial/internal/api/model
    instanceV1URLs:
        properties:
            streaming_api:
                description: Websockets address for status and notification streaming.
                example: wss://example.org
                type: string
                x-go-name: StreamingAPI
        title: InstanceV1URLs models instance-relevant URLs for client application consumption.
        type: object
        x-go-name: InstanceV1URLs
        x-go-package: github.com/superseriousbusiness/gotosocial/internal/api/model
...
host: example.org
info:
    contact:
        email: admin@gotosocial.org
        name: GoToSocial Authors
    license:
        name: AGPL3
        url: https://www.gnu.org/licenses/agpl-3.0.en.html
    title: GoToSocial Swagger documentation.
    version: REPLACE_ME
paths:
...
    /api/v1/instance:
        get:
            operationId: instanceGetV1
            produces:
                - application/json
            responses:
                "200":
                    description: Instance information.
                    schema:
                        $ref: '#/definitions/instanceV1'
                "406":
                    description: not acceptable
                "500":
                    description: internal error
            summary: View instance information.
            tags:
                - instance
        patch:
            consumes:
                - multipart/form-data
            description: This requires admin permissions on the instance.
            operationId: instanceUpdate
            parameters:
                - allowEmptyValue: true
                  description: Title to use for the instance.
                  in: formData
                  maximum: 40
                  name: title
                  type: string
                - allowEmptyValue: true
                  description: Username of the contact account. This must be the username of an instance admin.
                  in: formData
                  name: contact_username
                  type: string
                - allowEmptyValue: true
                  description: Email address to use as the instance contact.
                  in: formData
                  name: contact_email
                  type: string
                - allowEmptyValue: true
                  description: Short description of the instance.
                  in: formData
                  maximum: 500
                  name: short_description
                  type: string
                - allowEmptyValue: true
                  description: Longer description of the instance.
                  in: formData
                  maximum: 5000
                  name: description
                  type: string
                - allowEmptyValue: true
                  description: Terms and conditions of the instance.
                  in: formData
                  maximum: 5000
                  name: terms
                  type: string
                - description: Thumbnail image to use for the instance.
                  in: formData
                  name: thumbnail
                  type: file
                - description: Image description of the submitted instance thumbnail.
                  in: formData
                  name: thumbnail_description
                  type: string
                - description: Header image to use for the instance.
                  in: formData
                  name: header
                  type: file
            produces:
                - application/json
            responses:
                "200":
                    description: The newly updated instance.
                    schema:
                        $ref: '#/definitions/instance'
                "400":
                    description: bad request
                "401":
                    description: unauthorized
                "403":
                    description: forbidden
                "404":
                    description: not found
                "406":
                    description: not acceptable
                "500":
                    description: internal server error
            security:
                - OAuth2 Bearer:
                    - admin
            summary: Update your instance information and/or upload a new avatar/header for the instance.
            tags:
                - instance
...
schemes:
    - https
    - http
securityDefinitions:
    OAuth2 Application:
        flow: application
        scopes:
            write:accounts: grants write access to accounts
        tokenUrl: https://example.org/oauth/token
        type: oauth2
    OAuth2 Bearer:
        authorizationUrl: https://example.org/oauth/authorize
        flow: accessCode
        scopes:
            admin: grants admin access to everything
            admin:accounts: grants admin access to accounts
            read: grants read access to everything
            read:accounts: grants read access to accounts
            read:blocks: grant read access to blocks
            read:custom_emojis: grant read access to custom_emojis
            read:favourites: grant read access to favourites
            read:follows: grant read access to follows
            read:media: grant read access to media
            read:notifications: grants read access to notifications
            read:search: grant read access to searches
            read:statuses: grants read access to statuses
            read:streaming: grants read access to streaming api
            read:user: grants read access to user-level info
            write: grants write access to everything
            write:accounts: grants write access to accounts
            write:blocks: grants write access to blocks
            write:follows: grants write access to follows
            write:media: grants write access to media
            write:statuses: grants write access to statuses
            write:user: grants write access to user-level info
        tokenUrl: https://example.org/oauth/token
        type: oauth2
swagger: "2.0"